+++ src/app/add-user/add-user.component.ts
@@ -136,7 +136,7 @@ export class AddUserComponent implements OnInit {
       "email": this.addUserForm.get('email')?.value == "" ? this.addUserForm.get('address')?.value : undefined,
       "city": this.addUserForm.get('city')?.value == "" ? this.addUserForm.get('address')?.value : undefined,
       "state": this.addUserForm.get('state')?.value == "" ? this.addUserForm.get('address')?.value : undefined,
-      "new": this.addUserForm.get('new')?.value == "",
+      "new": this.addUserForm.get('new')?.value,
     };
     this.data = this.addUser;
   }