+++ src/app/highlight-search.pipe.ts
@@ -6,7 +6,8 @@ import { Pipe, PipeTransform } from '@angular/core';
 export class HighlightSearchPipe implements PipeTransform {
 
   transform(value: string, search: string): string {
-    const valueStr = value + ''; // Ensure numeric values are converted to strings
+    if(value == "" || value.length == 0) value = "---";
+    const valueStr = value + ''; // Ensure numeric values are converted to strings;
     return valueStr.replace(new RegExp('(?![^&;]+;)(?!<[^<>]*)(' + search + ')(?![^<>]*>)(?![^&;]+;)', 'gi'), '<strong class="your-class">$1</strong>');
   }
 }